Serve sections as linked documents from flat dir
This essentially turns the notebook into a wiki-like structure.

All files are saved in a flat directory structure (though this need not
be the case; we'll see how this goes). The ./build script creates
standalone html documents in the target directory, which ./serve will
serve with a python http server. Links work as expected.